@@ -171,8 +171,13 @@ namespace IdentityServer4.MicroService.Data
         /// 余额
         /// </summary>
         [Column(TypeName = "decimal(18,2)")]
-        public decimal Money { get; set; } 
+        public decimal Money { get; set; }
         #endregion
+
+        /// <summary>
+        /// 国家代码
+        /// </summary>
+        public string CountryCode { get; set; }
     }
 
     #region identity
@@ -376,6 +381,7 @@ namespace IdentityServer4.MicroService.Data
     {
         public long ParentUserID { get; set; }
         public string UserName { get; set; }
+        public string Avatar { get; set; }
         public System.DateTime CreateDate { get; set; }
         public bool LockFlag { get; set; }
         public long DataAmount { get; set; }
@@ -404,12 +410,17 @@ namespace IdentityServer4.MicroService.Data
         public string Pools { get; set; }
         public string PoolIDs { get; set; }
         public string Email { get; set; }
+        public bool EmailConfirmed { get; set; }
+        public string PhoneNumber { get; set; }
+        public bool PhoneNumberConfirmed { get; set; }
         public List<View_User_Role> Roles { get; set; }
         public List<View_User_Claim> Claims { get; set; }
         public List<View_User_File> Files { get; set; }
         public List<View_User_Property> Properties { get; set; }
         public List<View_User_Tenant> Tenants { get; set; }
-       
+
+        public List<View_User_Login> Logins { get; set; }
+
     }
     /// <summary>
     /// used for View_User Roles property
@@ -462,6 +473,16 @@ namespace IdentityServer4.MicroService.Data
     public class View_User_Tenant
     {
         public long TenantId { get; set; }
+    }
+
+    public class View_User_Login
+    {
+        public string LoginProvider { get; set; }
+        
+        public string ProviderDisplayName { get; set; }
+
+        public string ProviderKey { get; set; }
+
     }
     #endregion
 }
