@@ -5,12 +5,12 @@ namespace IdentityServer4.MicroService.Models.Apis.ApiResourceController
     public class ApiResourcePublishRequest
     {
         /// <summary>
-        /// 微服务名称
+        /// API名称
         /// </summary>
         public string name { get; set; }
 
         /// <summary>
-        /// 微服务功能简介
+        /// API功能简介
         /// </summary>
         public string description { get; set; }
 
